update for docker networking layer
This fixes issues with docker-compose containers killing the proxy. I'm
using a fork of docker-gen while I wait for the groupByTag PR to be
merged and released.

I've strayed further from the upstream nginx-proxy container now, I'm
not sure that it makes sense to build on top of it anymore. These
changes aren't likely to be accepted upstream since they're targetted
specifically at a local dev environment, which is not a primary use
case for nginx-proxy.