add CORS support
Conditionally adds CORS support to a container's nginx config
The condition is based on a container's CORS_ENABLED environment
variable. The default value is false, meaning that the existing
behavior will not change if the environment variable is missing.

e.g.

// docker-compose.yml
...
services:
  web:
    environment:
      - CORS_ENABLED=true

would enable CORS for that container's nginx config